requirements.txt
Flask
textblob
newspaper3k
flask_sqlalchemy
transformers
vaderSentiment
google-generativeai
python-dotenv
trafilatura